use ints instead of floats in checkers

+++ checkers/src/board.cpp
@@ -1,19 +1,19 @@
 #include "board.hpp"
 #include <utils/log.hpp>
 
-#include <iostream> // TODO: remove
+#include <iostream>  // TODO: remove
 
 namespace checkers {
 
 Board::Board(Rect position, TextureRef texture)
-    : GameObject{position, texture}
-    , _current_player{Color::WHITE}
-    , _current_selection{SelectionMode::SOURCE} {
+    : GameObject{position, texture},
+      _current_player{Color::WHITE},
+      _current_selection{SelectionMode::SOURCE} {
     decltype(auto) window = Window::instance();
     _cell_size = std::min(window.width(), window.height()) / _size;
     _components.reserve(_size * _size);
-    for (size_t i{}; i < _size; ++i) {
-        for (size_t j{}; j < _size; ++j) {
+    for (int i{}; i < _size; ++i) {
+        for (int j{}; j < _size; ++j) {
             auto position =
                 Rect{i * _cell_size, j * _cell_size, _cell_size, _cell_size};
             _components.push_back(std::make_shared<Cell>(position));
@@ -21,17 +21,17 @@ Board::Board(Rect position, TextureRef texture)
     }
 }
 
-bool Board::is_valid_position(size_t x, size_t y) const {
+bool Board::is_valid_position(int x, int y) const {
     // dark squares of the board are considered to be valid
     return y < _size && x < _size && ((x + y) % 2);
 }
 
-CellRef Board::at(size_t x, size_t y) {
+CellRef Board::at(int x, int y) {
     // TODO: understand why (x * _size + y) did not for diagonally !?!?!?
     return std::static_pointer_cast<Cell>(_components.at(y * _size + x));
 }
 
-size_t Board::size() {
+int Board::size() {
     return _size;
 }
 
@@ -43,8 +43,8 @@ void Board::handle_inputs() {
         return;
     }
     CellRef cell_on_focus{};
-    for (size_t i{}; i < _size; ++i) {
-        for (size_t j{}; j < _size; ++j) {
+    for (int i{}; i < _size; ++i) {
+        for (int j{}; j < _size; ++j) {
             auto current_cell = this->at(i, j);
             if (current_cell->mouse_hovers_over()) {
                 cell_on_focus = current_cell;
@@ -86,14 +86,17 @@ void Board::handle_inputs() {
 
 void Board::switch_players() {
     int current = static_cast<int>(_current_player);
-    Logger::instance().info("switching players : %d -> %d", current, 1 - current);
+    Logger::instance().info("switching players : %d -> %d", current,
+                            1 - current);
     _current_player = static_cast<Color>(1 - current);
 }
 
 void Board::switch_selection_modes() {
     int current = static_cast<int>(_current_selection);
-    Logger::instance().info("switching selection modes : %d -> %d", current, 1 - current);
-    _current_selection = static_cast<SelectionMode>(1 - static_cast<int>(_current_selection));
+    Logger::instance().info("switching selection modes : %d -> %d", current,
+                            1 - current);
+    _current_selection =
+        static_cast<SelectionMode>(1 - static_cast<int>(_current_selection));
 }
 
 void Board::move(CellRef src, CellRef dst) {
